API/requirements.txt
flask == 2.0.2
Flask-SQLAlchemy == 2.5.1
httpie == 2.6.0
python-dateutil == 2.8.2
psycopg2 == 2.9.1
Flask-Cors == 3.0.10